requirements.txt
chainlit
python-dotenv
openai
langfuse
openai-agents
qdrant-client
playwright
beautifulsoup4
Pillow
requests
nest-asyncio